refactor: reexport system core and update docs
