+++ ui_components.py
@@ -1,0 +1,6 @@
+"""DEPRECATED shim for backward compatibility.
+Use `from common.ui_components import ...` directly.
+This module will be removed in a future release.
+"""
+
+from common.ui_components import *  # noqa: F401,F403
